feat: add support for stateless function components
stateless components for react are a performance optimization we should be able to create proper
propTypes for it
