+++ package.json
@@ -57,6 +57,7 @@
     "typescript": "2.0.3"
   },
   "dependencies": {
+    "astq": "1.8.0",
     "babylon": "6.12.0",
     "dts-dom": "0.1.11",
     "minimist": "1.2.0"
