+++ tests/stateless.d.ts
@@ -1,0 +1,6 @@
+declare module 'component' {
+    export interface ComponentProps {
+        optionalAny?: any;
+    }
+    export function Component(props: ComponentProps): JSX.Element;
+}
