+++ tests/stateless.jsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+export const Component = ({optionalAny}) => <div />;
+Component.propTypes = {
+	optionalAny: React.PropTypes.any,
+};
