+++ typings/astq/astq.d.ts
@@ -1,0 +1,11 @@
+declare module 'astq' {
+
+  namespace astq {
+    class ASTQ {
+      public query(ast: any, query: string): any[];
+    }
+  }
+
+  export = astq;
+
+}
